+++ code/src/cljs/sixsq/nuvla/ui/apps_project/views.cljs
@@ -19,7 +19,6 @@
             [sixsq.nuvla.ui.utils.semantic-ui :as ui]
             [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]
             [sixsq.nuvla.ui.utils.style :as style]
-            [sixsq.nuvla.ui.utils.tab :as tab]
             [sixsq.nuvla.ui.utils.time :as time]
             [sixsq.nuvla.ui.utils.values :as values]))
 
@@ -129,7 +128,7 @@
                :centered  true}
       [ui/GridRow {:centered true}
        [ui/GridColumn
-        [apps-views-detail/OverviewDescription]]]
+        [apps-views-detail/OverviewDescription [::spec/tab]]]]
       [ui/GridRow
        [ui/GridColumn
         [ModulesView]]]