+++ code/src/cljs/sixsq/nuvla/ui/clouds/views.cljs
@@ -5,7 +5,7 @@
             [re-frame.db]
             [reagent.core :as r]
             [sixsq.nuvla.ui.acl.views :as acl]
-            [sixsq.nuvla.ui.clouds-detail.views :as infra-detail]
+            [sixsq.nuvla.ui.clouds-detail.views :as clouds-detail]
             [sixsq.nuvla.ui.clouds.events :as events]
             [sixsq.nuvla.ui.clouds.spec :as spec]
             [sixsq.nuvla.ui.clouds.subs :as subs]
@@ -47,7 +47,8 @@
        {:on-refresh #(dispatch [::events/get-infra-service-groups])}]]]))
 
 (defn ServiceCard
-  [{:keys [id name description path subtype logo-url swarm-enabled online] :as _service}]
+  [{:keys [id name description path subtype logo-url online]
+    :as infra-service}]
   (let [status (cond
                  (true? online) true
                  (false? online) false
@@ -63,16 +64,7 @@
                     (or name id)]
       :meta        path
       :description description
-      :content     (when (true? swarm-enabled)
-                     [ui/Label {:image    true
-                                :color    "blue"
-                                :circular true
-                                :basic    true
-                                :style    {:left   "0"
-                                           :margin "0.7em 0 0 0"}}
-                      [ui/Image {:bordered true}
-                       [ui/Icon {:name "docker"}]]
-                      "Swarm enabled"])}]))
+      :content     [clouds-detail/CompatibilityLabel infra-service]}]))
 
 (defn ServiceGroupCard
   [id name]
@@ -559,6 +551,6 @@
         root     [Infrastructures]
         children (case n
                    1 root
-                   2 [infra-detail/InfrastructureDetails uuid]
+                   2 [clouds-detail/InfrastructureDetails uuid]
                    root)]
     [ui/Segment style/basic children]))