+++ code/src/cljs/sixsq/nuvla/ui/main/effects.cljs
@@ -42,12 +42,14 @@
       (set-unload-protection)
       (clear-unload-protection))))
 
-(comment
-  (set-unload-protection)
-  (clear-unload-protection)
-  )
 
 (defn- stop-browser-back
+  "It's not possible to disable navigate back in modern browsers.
+
+   Workaround:
+   1. Push current route on history stack, so that navigate back leaves you on same page.
+   2. Navigating back then goes to same page, triggers 'popstate' event, which pushes current route again.
+       see: https://stackoverflow.com/a/64572567"
   [f]
   (.pushState js/window.history nil "" (.-href js/window.location))
   (set!
@@ -57,23 +59,34 @@
       (when (fn? f) (f)))))
 
 (defn- start-browser-back
-  []
+  [f nav-back?]
   (set! js/window.onpopstate nil)
-  (.back js/window.history))
+  ;; if modal was opened by navigating back, go back two steps in history stack
+  (cond
+    nav-back? (.go js/window.history -2)
+    (fn? f)   (f)
+    :else     (.back js/window.history)))
 
-(comment
-  (stop-browser-back +)
-  (start-browser-back)
-  )
 
 (reg-fx
   ::disable-browser-back
   (fn [f]
-    (js/console.error "disable-browser-back")
     (stop-browser-back f)))
+
 
 (reg-fx
   ::enable-browser-back
-  (fn [_]
-    (js/console.error "enable-browser-back effect")
-    (start-browser-back)))
+  (fn [{:keys [cb-fn nav-back?]}]
+    (start-browser-back cb-fn nav-back?)))
+
+(reg-fx
+  ::add-pop-state-listener-close-modal-event
+  (fn [f]
+    (.addEventListener js/window "popstate" f)))
+
+
+(reg-fx
+  ::clear-popstate-event-listener
+  (fn
+    [f]
+    (.removeEventListener js/window "popstate" f)))